Add integration test for schema shortcuts
As shortcuts for schema changes were introduced, a new integration
test is required to ensure schema integrity.

Test workflow is:

 - Enumerate all shortcuts and select target versions
 - Using incremental schema changes, apply each target version in
order and export the schema using `cqlsh`
 - For each shortcut directory, copy it over, apply target version,
call `cqlsh` to export the schema and compare it to the schema
generated from incremental changes. On success, remove the
directory.

Testing:
Tests pass with current schema

After modifying the shortcut schema manually, test fails with

```
Diff:
--- Expected
+++ Actual
@@ -518,3 +518,3 @@
     data blob,
-    data_encoding text,
+    data_encoding blob,
     data_version int,
```

as expected.